feat: Recognise the 'apple' provider ID in attribute importer mapper
This allows the provider to be configure `keycloak_attribute_importer_identity_provider_mapper` resources with the [Apple Identity Provider for Keycloak plug-in](https://github.com/klausbetz/apple-identity-provider-keycloak).

Fortunately, all the actual differences are buried in the provider itself and the mapper just needs to do the same thing as for `facebook` and `google`.

Signed-off-by: Graham Reed <greed@hypervolt.co.uk>