Do not delete the "account" client in the acceptance test because it is undeletable (KeyCloak returns http status 400)
Terraform acceptance test library simply can't handle this situation. One way that I found to overcome this is to replace the `DeleteContext` function in acceptance tests only.

I believe we must not alter the behavior of the standard response of KeyCloak when a client is deleted, nor hide that fact that KeyCloak response is 400 in the production code.

Signed-off-by: Mikhail Putilov <Mikhail.Putilov@dinape.com>